pyproject.toml
[project]
name = "cxc-framework"
version = "0.1.0"
description = "Cortex Code (CxC) Framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "GitPython>=3.0.0",
    "schedule>=1.2.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "aiosqlite>=0.19.0",
    "boto3>=1.26.0",
    "rich>=13.0.0",
    "mcp[cli]>=1.0.0",
]

[project.scripts]
cxc = "cxc.cli:main"
cxc-mcp = "cxc.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cxc", "commands", "templates"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.10",
]


